refactor code for swapping and shuffling - add isSolved to redux state
