+++ .eslintrc.json
@@ -8,6 +8,7 @@
         "object-curly-newline": "off",
         "no-underscore-dangle": "off",
         "import/prefer-default-export": "off",
-        "react/destructuring-assignment": "off"
+        "react/destructuring-assignment": "off",
+        "no-plusplus": "off"
     }
 }
