+++ client/src/constants.js
@@ -1,3 +1,5 @@
 export const SHUFFLE = 'SHUFFLE';
 export const SWAP = 'SWAP';
 export const EMPTY_TILE = 15;
+export const NUM_OF_ROWS = 4;
+export const NUM_OF_COLS = 4;
