+++ client/src/reducers/initialState.js
@@ -1,11 +1,13 @@
-import { range, shuffle, without } from 'lodash';
+import { range } from 'lodash';
 
-import { EMPTY_TILE } from '../constants';
+import { shuffleNumbers } from '../utils';
 
 const numbers = range(0, 16);
 
 export default {
   puzzleState: {
-    puzzleNumbers: shuffle(without(numbers, EMPTY_TILE)).concat(EMPTY_TILE),
+    puzzleNumbers: shuffleNumbers(numbers),
+    isSolved: false,
+    sortedNumbers: range(0, 16),
   },
 };
