+++ client/src/utils/index.js
@@ -1,34 +1,4 @@
-import { EMPTY_TILE } from '../constants';
+import { swap } from './swap';
+import { shuffleNumbers } from './shuffle';
 
-export const NUM_OF_ROWS = 4;
-export const NUM_OF_COLS = 4;
-
-export const getMatrixPosition = (index, cols) => ({
-  row: Math.floor(index / cols),
-  col: index % cols,
-});
-
-const canSwap = (gridIndex, emptyTileIndex) => {
-  const { row: srcRow, col: srcCol } = getMatrixPosition(
-    gridIndex,
-    NUM_OF_COLS,
-  );
-  const { row: destRow, col: destCol } = getMatrixPosition(
-    emptyTileIndex,
-    NUM_OF_COLS,
-  );
-  return Math.abs(srcRow - destRow) + Math.abs(srcCol - destCol) === 1;
-};
-
-export const swap = (puzzleNumbers, gridIndex) => {
-  const emptyTileIndex = puzzleNumbers.indexOf(EMPTY_TILE);
-  if (canSwap(gridIndex, emptyTileIndex)) {
-    const copyArr = [...puzzleNumbers];
-    [copyArr[gridIndex], copyArr[emptyTileIndex]] = [
-      copyArr[emptyTileIndex],
-      copyArr[gridIndex],
-    ];
-    return copyArr;
-  }
-  return puzzleNumbers;
-};
+export { swap, shuffleNumbers };
